Drop issuegen and motdgen .path units
The intention was to have the issue/motd update whenever a file
was dropped in or modified in the private issue/motd directories,
however the path units are not working for this case. The motd/issue
will be updated on reboot if the service unit is enabled - there
isn't a need to respond to the changes immediately.