issuegen: do not share the staged file
With the `staged` file shared, there was potential for two or
more processes executing `issuegen` to write to it, resulting
in corrupted output, or the error in the `cat` command due to
missing file reported in https://github.com/coreos/console-login-helper-messages/issues/35#issue-593088209.

Instead, write the intermediate output to a variable before writing
to the final output file. This ensures only valid output is written
to the issue file shown to the terminal.

If large issue snippets are ever passed, a tempfile is another
option for storing intermediate output to avoid storing snippets
in memory. However, practically, snippet sizes should be small, as
they are intended to be read on a terminal.