issuegen: do not use a staged file
Avoid using an intermediate staged file to write new issue snippets
to. Previously, I had added the `staged` file so that there would be
a temporary "scratch" pad to write the new issue snippets to before
writing to the final location (`generated`). This would mean that
only completed writes to `generated` would be made, by copying from
`staged`.

However, the `staged` file itself is not giving proper control of
concurrency of reads or writes to the `generated` location. Since
writes to the file are completed by a system call to the OS, the
OS will ensure that the `generated` file is valid when read. To
ensure that the `generated` file is read at the proper time (i.e.
when the `issuegen` script has completely finished) by other
system services or users, systemd units should be used for this.

In removing the `staged` file, the error in the `cat` command
reported in https://github.com/rfairley/console-login-helper-messages/issues/35#issue-593088209
is avoided. Instead of exiting the script with error when
concurrent writes are made, there will simply be two writes of the
same content to the output file.