motdgen: do not use a staged file
For the same reason as in https://github.com/rfairley/console-login-helper-messages/pull/36
(copying the phrasing below):

Avoid using an intermediate staged file to write new motd snippets
to. Previously, I had added the `staged` file so that there would be
a temporary "scratch" pad to write the new motd snippets to before
writing to the final location (`generated`). This would mean that
only completed writes to `generated` would be made, by copying from
`staged`.

However, the `staged` file itself is not giving proper control of
concurrency of reads or writes to the `generated` location. Since
writes to the file are completed by a system call to the OS, the
OS will ensure that the `generated` file is valid when read. To
ensure that the `generated` file is read at the proper time (i.e.
when the `motdgen` script has completely finished) by other
system services or users, systemd units should be used for this.

Additionally, perform code tidyups similar to those done for
`issuegen` in https://github.com/rfairley/console-login-helper-messages/pull/37.