issuegen: set output paths based on util-linux version
If the util-linux version installed is greater than or equal to
2.35, when support for /run/issue.d was added, set the the output
paths so that /run/issue.d is utilized.

The version check for util-linux is done by using `rpm` to check
the version of an installed util-linux RPM package. For
non-RPM-based systems, the output paths will take a default
value of the private directory location /run/console-login-helper-messages/issue.d.
`rpm` is chosen to perform the check as the only kown consumers of
console-login-helper-messages are RPM-based distributions, however
support could always later be added for Debian-based or other
Linux distributions.