issuegen: add path unit
Add `console-login-helper-messages-issuegen.path`, so that
`issuegen` is run whenever snippets are written to `/run/console-login-helper-messages/issue.d`.
This should be the main way of triggering a run of `issuegen` (first
trigger the path, which starts the service, which will run
`issuegen`). The only other time the `issuegen` service is invoked
is on boot (if the unit is enabled).

Enablement of the path unit is required upon installation, in order
for this to work correctly.