motdgen: add path unit
Add `console-login-helper-messages-motdgen.path`, so that
`motdgen` is run whenever snippets are written to `/run/console-login-helper-messages/motd.d`.
This should be the main way of triggering a run of `motdgen` (first
trigger the path, which starts the service, which will run
`motdgen`). The only other time the `motdgen` service is invoked
is on boot (if the unit is enabled).

Note rate limiting of starts on the .service units may be a conern,
but the defaults (at least in a VM running Fedora) appear to be
around 10,000 starts after dropping snippets as fast as possible
through a while loop into the directories watched by PathChanged,
which is more than enough for most use cases. The path unit should
be reliable enough without configuring any rate limiting, but
`StartLimitIntervalSec=0` can always be set in the unit if it does
become a problem. A clear error message in systemd
`start-limit-hit` should be given, if it is a problem.

Enablement of the path unit is required upon installation, in order
for this to work correctly.

Note that Container Linux originally used a path unit in a similar
way for responding to changes to a specific file:
https://github.com/coreos/init/blob/master/systemd/system/motdgen.path