issuegen.defs: move under usr/lib
This file need not be executable as it is only sourced into
scripts and not directly executed. Move out of `usr/libexec` and
into `usr/lib`.

Update permission bits to be non-executable, and while at it
fix the shebang to use the full path (as RPM will mangle the
symlinked path to use the fully qualified path).

also unset its executable bit, as this file is only sourced, and
remove set bash flags and fix its shebang