issuegen,motdgen: move systemd-user-sessions dependency to path unit
With the .path units enabled, the .path units trigger the issuegen
and motdgen services during startup and throughout the system runtime,
upon issue or motd snippets being dropped into runtime directories.
Make sure the .path units run before `systemd-user-sessions.service`,
rather than the .service units. With this, there is now no requirement
to have both the .service and .path units for issuegen and motdgen
enabled; only the .path need to be enabled to ensure the final
issue/motd messages are generated before the issue/motd are shown to
the user.