add script to sync source code to VM
Add a script to execute the `make install` target to a local
directory via a container, and rsync the directory to the root
of the virtual machine. This avoids the extra work in building an
RPM with `install_vm_rpms.sh`, and means a Makefile versioned with
the source can be used rather than updating a specfile which
comes from a separate git repository (Fedora SCM).