gensnippet: Support complex networking devices
NetworkManager dispatcher script:
Add NetworkManager dispatcher script that triggers gensnippet_if when a new
interface has been activated, instead of using udev rules. A dispatcher
script is more suitable over a permissive udev rule because it allows us to
filter out non-real devices by checking whether a new interface is a real device.
This should allow console-login-helper-messages to support complex networking
devices, as well as networking devices with custom names, as we are no longer
reliant on devices having names starting with 'e'.

gensnippet_if:
Generate new issue snippet if device is real. If device is a bond, team, bridge,
etc., generate new issue snippet if backed by any real devices. Interfaces that
are part of a bond,team, etc. are still displayed, despite not having any IP
address information because it might be helpful for users to know all the
interfaces that are connected, regardless of their membership in bonds or teams.