Makefile: add symlink
Add a symlink to link the concatentated issue file to a directory
that agetty reads from (i.e. `/etc/issue.d`). This was originally
done in the Specfile, but since we are unifying the Makefile and
Specfile to avoid multiple sources of truth, we let the `install`
target handle everything installation-related.