tests/kola/basic/test-issue.sh: Update expected symlink
There was a discrepancy between the exact path of the symlink that
links `/etc/issue.d/40_console-login-helper-messages.issue` to the
generated `40_console-login-helper-messages.issue` between the
Makefile and `tmpfiles.d` conf file. Update the expected symlink
to be the one in the `tmpfiles.d` conf file since we no longer
create the symlink in the Makefile.