{issuegen,motdgen}.{service,path}: Remove `PathExistsGlob` and add dependency on `.path`
A recent change in systemd version 246 (which will be in f33)
https://github.com/systemd/systemd/commit/708961c7011b7578d3e6e10ca463d15edb62b5c1
changed the behaviour of `PathExistsGlob` option in `.path` units;
we can no longer use it for our original purpose.

Remove this option and also add `After=` dependencies on `.path`
units in the `.service` units to ensure that we have similar behaviour
as before. Presets of `.service` units are now required since they
must be started at least once during boot up.