gensnippet_if: Fix infinite recursive function
Due to bash leaving non-matching wildcard patterns unchanged
(instead of expanding it to an empty list), if a device is not real
and does not have a subordinate device, `is_iface_backed_by_real_device()`
will get stuck in an infinite recursive loop, causing segmentation
faults.

Ignore directories that are the result of non-matching wildcard
patterns.