issuegen.service: Remove `Before=` dependency on `user-sessions.service`
It should not matter whether `issuegen.service` is run before
`systemd-user-sessions.service` or not. As long as it is run any
time after `issuegen.path`, all issue files in the private `issue.d`
directories should be detected and concatenated. So for consistency
with `motdgen.service`, remove this since it already doesn't exist
there.