gensnippet-ssh-keys: Ensure service runs before agetty
During the major rearchitecting to switch CLHM to use directories
of upstream projects (util-linux), we accidentally dropped a
`Before=` dependency on `systemd-user-sessions.service` when
generating ssh keys snippets. This was previously done in
`issuegen.service`, but that service unit no longer exists, so we
now add the ordering dependency directly in
`gensnippet-ssh-keys.service` to ensure that we drop the files in
the necessary locations before `agetty` starts.

We also don't need the extra `agetty --reload` since we are only
writing files once, and it happens before `agetty` is started.
(This was previously required because `issuegen` supported users
dynamically dropping snippets into CLHM's private directories).